Fix IBC transfers when using router
There is an issue that if one of the cascaded calls in the router fails, the whole packet is considered failed, but since we don't rollback the previous changes (e.g. ics-20 transfer), the packet is state becomes invalid and may cause some weird issues, e.g. allowing to submit the same packet multiple times, receiving the more tokens each time (double spending).

It's currently not clear whether it's possible to allow other modules to fail in case we have pre-checked all the sub-calls in the beginning, because then there is a possibility to mark a good packet (from ics-20 module's point of view) as failed, although, it shouldn't fail in reality, because initially, the process of fee collection and memo execution was not defined in the standard.