feat: Add eth_query_proof_2d function
This commit adds a new function called eth_query_proof_2d to the Client struct. This function takes two keys, a block height, and a storage index as input parameters and returns an EIP1186ProofResponse. The implementation of this function uses ethers-rs library to encode the keys and get proof from the client.