Refactor Makefile and yui-clone.sh, delete yui-light-client-abi.sh, and modify yui-storage.sh.
The Makefile in hyperspace was refactored to remove a script that generated contract storage layout. The yui-clone.sh script was modified to use environment variables for the IBC path and repository. The script also checks if the directory exists before cloning the repo. The yui-light-client-abi.sh script was deleted as it is no longer needed. Finally, the forge inspect command in yui-storage.sh now takes an argument for which contract's storage layout to inspect instead of always using OwnableIBCHandler.