Refactor subxt types to string from MemoMessage
The changes involve removing the custom `MemoMessage` type in favor of the simpler and more generic `String` type. This refactoring makes the code more flexible and easier to understand. Another change is the replacement of the `PicassoBeefyAuthoritySet` with a placeholder `DummyBeefyAuthoritySet.` This is assumed to be a temporary change, made for the purpose of debugging or isolating a problem. These changes were introduced in both `picasso_rococo.rs` and `composable.rs` files.