Fix various clippy and compilation warnings (#442)
In my previous commit I’ve introduced a new build warning.  Motivation for
this commit was to fix it and while I’m at it address some other issues as
well.  However, turns out there’s too many warnings I care to deal with so
this addresses only some of the warnings.