audit-4: client never expires (#92)
Implement expired() to set the client to expire if the next update
is after the 27 hours of a sync committee change.

There's a trade-off here right now. As a higher number gives us
margin to not update the state so often. However, it needs to be
lower than the unbonding period (which is 28 hours)